// src/app/model/Employee.ts
export class Employee{

	apellidoPaterno?: string;
	apellidoMaterno?: string;
	correo?: string;
	nc?: string;
	nombre?: string;
	nombreCompleto?: string;
	nss?: string;
	celular?: string;
	telefono?:string;
	lada?:string;
	dia?: string;
	mes?: string;
	anio?: string;
	meses?: any;
}